requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
python-dotenv
pytest